Canon EDSDK update

* Change Gain* from short to int to accomodate high ISO speeds

* Bring Canon driver up to speed with current EDSDK

* Fix up comment

* Add release notes

* Update release notes

* More gain-related short -> int


Former-commit-id: 60b4e02f42d9d777feceadc09fb410b9b7dbe620
Former-commit-id: 42f8f3a5f1ef6457834e54e0810e91d8f5f6870c

+++ NINA/Model/MyCamera/ASICamera.cs
@@ -502,9 +502,9 @@ namespace NINA.Model.MyCamera {
             }
         }
 
-        public short Gain {
+        public int Gain {
             get {
-                return (short)GetControlValue(ASICameraDll.ASI_CONTROL_TYPE.ASI_GAIN);
+                return GetControlValue(ASICameraDll.ASI_CONTROL_TYPE.ASI_GAIN);
             }
             set {
                 if (SetControlValue(ASICameraDll.ASI_CONTROL_TYPE.ASI_GAIN, value)) {
@@ -513,15 +513,15 @@ namespace NINA.Model.MyCamera {
             }
         }
 
-        public short GainMax {
+        public int GainMax {
             get {
-                return (short)GetControlMaxValue(ASICameraDll.ASI_CONTROL_TYPE.ASI_GAIN);
+                return GetControlMaxValue(ASICameraDll.ASI_CONTROL_TYPE.ASI_GAIN);
             }
         }
 
-        public short GainMin {
+        public int GainMin {
             get {
-                return (short)GetControlMinValue(ASICameraDll.ASI_CONTROL_TYPE.ASI_GAIN);
+                return GetControlMinValue(ASICameraDll.ASI_CONTROL_TYPE.ASI_GAIN);
             }
         }
