+++ NINA/ViewModel/Equipment/Camera/ICameraVM.cs
@@ -34,7 +34,7 @@ namespace NINA.ViewModel.Equipment.Camera {
 
         void SetBinning(short x, short y);
 
-        void SetGain(short gain);
+        void SetGain(int gain);
 
         void SetSubSample(bool subSample);
